+++ server/component/style/book/BookView.php
@@ -17,6 +17,18 @@ class BookView extends StyleView
     // Config field holidng infor for turn.js
     private $config;
 
+    /**
+     * DB field 'label_next' (empty string).
+     * The label of the navigation button to go to the next item.
+     */
+    private $label_next;
+
+    /**
+     * DB field 'label_back' (empty string).
+     * The label of the navigation button to go to the pervious item.
+     */
+    private $label_back;
+
     /* Constructors ***********************************************************/
 
     /**
@@ -30,6 +42,8 @@ class BookView extends StyleView
         parent::__construct($model);
         $this->title = $this->model->get_db_field("title");
         $this->config = $this->model->get_db_field("config");
+        $this->label_next = $this->model->get_db_field("label_next");
+        $this->label_back = $this->model->get_db_field("label_back");
     }
 
     /* Private Methods ********************************************************/