+++ CHANGELOG.md
@@ -9,6 +9,9 @@
  - add field `locked_after_submit` in style `slider`. If selected and if the field is used in a form that is not `is_log`, once the value is set, the field will not be able to be edited anymore.
  - add field `locked_after_submit` in style `textarea`. If selected and if the field is used in a form that is not `is_log`, once the value is set, the field will not be able to be edited anymore.
 
+### Bugfix
+ - in style `radio` the field `is_required` is proeprly checked
+
 # v4.1.2
 
 ### New Features